+++ src/helpers/chainIdToChain.ts
@@ -1,9 +1,18 @@
 import { Chain } from "viem";
-import { base, baseSepolia } from "viem/chains";
+import {
+  base,
+  baseSepolia,
+  mainnet,
+  robinhood,
+  unichain,
+} from "viem/chains";
 
 export const chainIdToChain: {
   [key: number]: Chain;
 } = {
   [base.id]: base,
   [baseSepolia.id]: baseSepolia,
+  [mainnet.id]: mainnet,
+  [unichain.id]: unichain,
+  [robinhood.id]: robinhood,
 };