+++ src/helpers/index.ts
@@ -1,4 +1,5 @@
 export * from "./hex";
 export * from "./ipfs";
 export * from "./chainIdToChain";
+export { isChainSupported, doesChainSupportSplitManager } from "./supportedChains";
 export * from "./permissions";